requirements.txt
# Core
python-dotenv>=1.0.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
aiohttp>=3.9.5
pyjwt>=2.8.0
openai>=1.30.0

# LiveKit Agents v1.x
livekit-agents>=1.5.0
livekit-api>=1.0.0
livekit-plugins-silero>=1.5.0
livekit-plugins-cartesia>=1.5.0
livekit-plugins-openai>=1.5.0

# STT and LLM via Groq (used as OpenAI-compatible client)
groq>=0.10.0

# RAG / Vector DB
chromadb>=0.5.3
sentence-transformers>=3.0.1

# Audio processing & conversion
pydub>=0.25.1
scipy>=1.13.1
numpy>=1.24.0